requirements.txt
# requirements.txt
paramiko
streamlit
pytest
flake8
matplotlib
scipy
plotly
black
isort
PyYAML
dotenv
psycopg2-binary>=2.9
# black, isort, mypy, etc.,